+++ src/overlaybd/zfile/compressor.cpp
@@ -91,7 +91,7 @@ public:
         }
         off_t src_offset = 0, dst_offset = 0;
         int ret = 0;
-        for (ssize_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < n; i++) {
             uncompressed_data[i] = ((unsigned char *)src + src_offset);
             compressed_data[i] = ((unsigned char *)dst + dst_offset);
             src_offset += src_chunk_len[i];
@@ -120,7 +120,7 @@ public:
         }
         off_t src_offset = 0, dst_offset = 0;
         int ret = 0;
-        for (ssize_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < n; i++) {
             compressed_data[i] = ((unsigned char *)src + src_offset);
             uncompressed_data[i] = ((unsigned char *)dst + dst_offset);
             src_offset += src_chunk_len[i];
@@ -208,7 +208,7 @@ public:
             return ret;
         }
 #endif
-        for (ssize_t i = 0; i < nblock; i++) {
+        for (size_t i = 0; i < nblock; i++) {
             ret =
                 LZ4_compress_default((const char *)uncompressed_data[i], (char *)compressed_data[i],
                                      src_chunk_len[i], dst_buffer_capacity / nblock);
@@ -240,7 +240,7 @@ public:
             return ret;
         }
 #endif
-        for (ssize_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < n; i++) {
             ret =
                 LZ4_decompress_safe((const char *)compressed_data[i], (char *)uncompressed_data[i],
                                     src_chunk_len[i], dst_buffer_capacity / n);
@@ -295,7 +295,7 @@ public:
                             size_t nblock) override {
 
         int ret = 0;
-        for (int i = 0; i < nblock; i++) {
+        for (size_t i = 0; i < nblock; i++) {
             ret = compress(uncompressed_data[i], src_chunk_len[i], compressed_data[i],
                            dst_buffer_capacity / nblock);
             if (ret < 0) {
@@ -311,7 +311,7 @@ public:
                               size_t dst_buffer_capacity, size_t nblock) override {
 
         int ret = 0;
-        for (ssize_t i = 0; i < nblock; i++) {
+        for (size_t i = 0; i < nblock; i++) {
             ret = decompress((const unsigned char *)uncompressed_data[i], src_chunk_len[i],
                              compressed_data[i], dst_buffer_capacity / nblock);
